+++ src/scheduler/scheduler.rs
@@ -90,7 +90,7 @@ impl Schedulable for Task {
 
     fn state(&self) -> TaskState {
         unsafe {
-            core::mem::transmute(self.state.load(Ordering::SeqCst))
+            core::mem::transmute(self.state.load(Ordering::SeqCst) as u32)
         }
     }
